R12:G5 温控 PI+前馈与电机-加热联锁——软件安全核心落地,-Werror 抓出 110k RPM 域宽真 bug
platform/control/thermal_pi.c:PI+风量前馈+条件积分抗饱和+硬顶(≥115°C 强制 0);
platform/safety/interlock.c:OK/DERATE/CUTOFF,风量丧失持续超时切断、闩锁、
恢复需「风量+冷却」双条件、无效读数即切(US4003388A 语义);
经 board ops(heater_permille,TRIAC 过零调功落位)接入 products 50Hz 任务。
host 单测 16/16(含一阶炉温仿真:收敛±3°C/过冲≤5°C),HOST_TEST_MIN 9→25;
bench 增 case-003(恢复缺冷却确认),run_bench 泛化每题自带 ref/test/includes,
BUGFIX_MIN 2→3,重签哈希锁。
域宽修复:speed_rpm uint16→uint32(110,000 RPM 在 uint16 回绕成 44464 会骗过联锁,
-Werror 编译期抓出;protection.h/interlock/main 全链同步)。

+++ swe-agent/firmware/platform/safety/protection.h
@@ -31,7 +31,7 @@ typedef struct {
     bool     valid;        /* 传感器读数有效性;false → 直接 TRIP */
     int16_t  temp_c;       /* 功率级温度 */
     uint16_t current_ma;   /* 母线电流 */
-    uint16_t speed_rpm;    /* 转速(0 = 停转) */
+    uint32_t speed_rpm;    /* 转速(0 = 停转;110k RPM 级,uint16 会回绕 — 域宽 bug 已被 -Werror 抓出) */
     uint16_t dt_ms;        /* 距上次调用的毫秒数 */
 } prot_inputs_t;
